Moves files to api/components/reducers/style dirs.
